execute the commands instead of begging the user to do it
It should be relatively easy to offer the option to execute the commands, instead of begging the user to do it. I for one generate my private keys with the same account that deals with letsencrypt.org, which technically doesn't protect private keys with unix file permissons, but it's convenient at least. It's nicer anyway, to have a standard syntax for lists of commands anyway, instead of just sticking them in the string and trying to figure out which argument {4} corresponds to.