added listen-port option for temp python server
Since port 80 is a privileged port, regular applications cannot bind to
it (that is, listen to / read traffic from it) by default under Linux.
Rather than granting special rights to allow this, some servers simply
route traffic on to other ports (such as 8080) after initially being
received on port 80.

This change allows the listening port used in the non-file-based mode to
be configured as needed.